+++ src/main/java/uk/gov/hmcts/reform/pip/channel/management/services/filegeneration/helpers/listmanipulation/CrownWarnedListManipulation.java
@@ -1,12 +1,12 @@
 package uk.gov.hmcts.reform.pip.channel.management.services.filegeneration.helpers.listmanipulation;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.node.ObjectNode;
 import uk.gov.hmcts.reform.pip.channel.management.models.external.datamanagement.Language;
 import uk.gov.hmcts.reform.pip.channel.management.models.templatemodels.CrownWarnedList;
 import uk.gov.hmcts.reform.pip.channel.management.services.filegeneration.helpers.CaseHelper;
 import uk.gov.hmcts.reform.pip.channel.management.services.filegeneration.helpers.DateHelper;
 import uk.gov.hmcts.reform.pip.channel.management.services.filegeneration.helpers.GeneralHelper;
+import uk.gov.hmcts.reform.pip.channel.management.services.filegeneration.helpers.PartyRoleHelper;
 
 import java.util.ArrayList;
 import java.util.Comparator;
@@ -16,7 +16,6 @@ import java.util.Map;
 import java.util.stream.Collectors;
 
 public final class CrownWarnedListManipulation {
-    private static final String DELIMITER = ", ";
     private static final String DEFENDANT = "defendant";
     private static final String DEFENDANT_REPRESENTATIVE = "defendantRepresentative";
     private static final String PROSECUTING_AUTHORITY = "prosecutingAuthority";
@@ -46,7 +45,7 @@ public final class CrownWarnedListManipulation {
                                                                              "dd/MM/yyyy");
                         sitting.get("hearing").forEach(hearing -> {
                             String listNote = GeneralHelper.findAndReturnNodeText(hearing, "listNote");
-                            handleParties(hearing);
+                            PartyRoleHelper.handleParties(hearing);
                             List<CrownWarnedList> rows = new ArrayList<>();
                             hearing.get("case").forEach(hearingCase -> {
                                 CaseHelper.formatLinkedCases(hearingCase);
@@ -69,48 +68,6 @@ public final class CrownWarnedListManipulation {
         });
 
         return sort(result);
-    }
-
-    static void handleParties(JsonNode hearing) {
-        List<String> defendants = new ArrayList<>();
-        List<String> defendantRepresentatives = new ArrayList<>();
-        List<String> prosecutingAuthorities = new ArrayList<>();
-
-        hearing.get("party").forEach(party -> {
-            if (!GeneralHelper.findAndReturnNodeText(party, "partyRole").isEmpty()) {
-                switch (party.get("partyRole").asText()) {
-                    case "DEFENDANT":
-                        defendants.add(createIndividualDetails(party));
-                        break;
-                    case "DEFENDANT_REPRESENTATIVE":
-                        defendantRepresentatives.add(createOrganisationDetails(party));
-                        break;
-                    case "PROSECUTING_AUTHORITY":
-                        prosecutingAuthorities.add(createOrganisationDetails(party));
-                        break;
-                    default:
-                        break;
-                }
-            }
-        });
-
-        ((ObjectNode) hearing).put(DEFENDANT, String.join(DELIMITER, defendants));
-        ((ObjectNode) hearing).put(DEFENDANT_REPRESENTATIVE, String.join(DELIMITER, defendantRepresentatives));
-        ((ObjectNode) hearing).put(PROSECUTING_AUTHORITY, String.join(DELIMITER, prosecutingAuthorities));
-    }
-
-    private static String createIndividualDetails(JsonNode party) {
-        JsonNode individualDetails = party.get("individualDetails");
-        String forenames = GeneralHelper.findAndReturnNodeText(individualDetails, "individualForenames");
-        String surname = GeneralHelper.findAndReturnNodeText(individualDetails, "individualSurname");
-
-        return surname + (surname.isEmpty() || forenames.isEmpty() ? "" : ", ")
-            + forenames;
-    }
-
-    private static String createOrganisationDetails(JsonNode party) {
-        JsonNode organisationDetails = party.get("organisationDetails");
-        return GeneralHelper.findAndReturnNodeText(organisationDetails, "organisationName");
     }
 
     private static Map<String, List<CrownWarnedList>> sort(Map<String, List<CrownWarnedList>> cases) {
