+++ src/main/resources/templates/magistratesPublicList.html
@@ -236,7 +236,7 @@
                           <td class="govuk-table__cell" th:classappend="${case.get('bottomBorder').asText()}" th:text="${case.get('caseNumber').asText()}"/>
                           <td class="govuk-table__cell" th:classappend="${case.get('bottomBorder').asText()}" th:text="${hearing.get('defendant').asText()}"/>
                           <td class="govuk-table__cell no-wrap" th:classappend="${case.get('bottomBorder').asText()}" th:text="${hearing.get('hearingType').asText()}"/>
-                          <td class="govuk-table__cell" th:classappend="${case.get('bottomBorder').asText()}" th:text="${hearing.get('prosecuting_authority').asText()}"/>
+                          <td class="govuk-table__cell" th:classappend="${case.get('bottomBorder').asText()}" th:text="${hearing.get('prosecutingAuthority').asText()}"/>
                           <td class="govuk-table__cell no-wrap" th:classappend="${case.get('bottomBorder').asText()}" th:text="${sitting.get('formattedDuration').asText() + ' ' + case.get('caseSequenceIndicator').asText()}"/>
                         </tr>
                         <span th:if="${hearing.get('listingNotes').asText() != ''}">
